requirements_pi.txt
absl-py==0.14.0
aioice==0.7.5
arandr==0.1.10
asn1crypto==0.24.0
astroid==2.5.1
asttokens==2.0.4
astunparse==1.6.3
automationhat==0.2.0
beautifulsoup4==4.9.3
bidict==0.21.3
blinker==1.4
blinkt==0.1.2
bottle==0.12.19
buttonshim==0.0.2
cachetools==4.2.2
Cap1xxx==0.1.3
certifi==2020.6.20
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==7.1.2
colorama==0.4.4
colorzero==1.1
cookies==2.2.1
crc32c==2.2.post0
cryptography==3.3.2
cupshelpers==1.0
Cython==0.29.24
dbus-python==1.2.16
decorator==4.4.2
Deprecated==1.2.13
distro==1.5.0
dnspython==2.1.0
docutils==0.16
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
eventlet==0.32.0
ExplorerHAT==0.4.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==4.3.1
fourletterphat==0.1.0
gast==0.5.2
gcloud==0.18.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpiozero==1.6.2
greenlet==1.1.1
grpcio==1.40.0
gyp==0.1
html5lib==1.1
httplib2==0.19.1
idna==2.10
ikpy==3.0.1
importlib-metadata==4.8.1
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==2.11.3
jwcrypto==1.0
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==0.0.0
logilab-common==1.8.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mpmath==1.2.1
mypy==0.812
mypy-extensions==0.4.3
netifaces==0.11.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.3.56
opt-einsum==3.3.0
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
poppy-creature==2.0.0
prettytable==2.2.0
protobuf==4.0.0rc2
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.1
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pyee==8.2.2
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypot==3.1.3
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1
python-engineio==3.13.2
python-jwt==3.3.0
python-socketio==4.6.0
pyxdg==0.25
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
requests-toolbelt==0.9.1
responses==0.12.1
roman==2.0.0
RPi.GPIO==0.7.0
rsa==4.7.2
RTIMULib==7.2.1
scipy==1.6.0
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.6.0b1
sense-hat==2.2.0
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==2.2.1
spidev==3.5
sseclient==0.0.27
ssh-import-id==5.10
sympy==1.8
termcolor==1.1.0
thonny==3.3.14
toml==0.10.1
tornado==6.1
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
unicornhathd==0.0.4
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
zipp==3.5.0
